requirements.txt
Django==4.2.8
python-dotenv~=1.0.0
python-telegram-bot==13.7